% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{all_combinationsCoNI}
\alias{all_combinationsCoNI}
\title{Robustness function}
\usage{
all_combinationsCoNI(
  vertexD,
  edgeD,
  no_samplesOut = 2,
  edgeDname = "edgeD",
  vertexDname = "vertexD",
  outputName = "CoNIOutput",
  outputDir = "./CoNIOutput",
  padjustvertexD = TRUE,
  correlateDFs = TRUE,
  splitedgeD = TRUE,
  split_number = 2,
  delPrevious = FALSE,
  delIntermediaryFiles = TRUE
)
}
\description{
This function takes as input the corresponding vertex Data and edge Data. Sample names for both datasets should match (samples are in the rows). User has to provide the number of samples that will be excluded (1,2, or n samples). Function will iteratively exclude all combinations for the number provided. Default value is 2. For very big datasets might fail.
}
\keyword{"no_samplesOut",}
\keyword{CoNI}
\keyword{It}
\keyword{a}
\keyword{an}
\keyword{as}
\keyword{available.}
\keyword{be}
\keyword{computer}
\keyword{enough}
\keyword{excluded.}
\keyword{experimental}
\keyword{extra}
\keyword{if}
\keyword{is}
\keyword{lot}
\keyword{not}
\keyword{number}
\keyword{of}
\keyword{parameter}
\keyword{parameters}
\keyword{plus}
\keyword{power}
\keyword{requires}
\keyword{run}
\keyword{same}
\keyword{samples}
\keyword{specifies}
\keyword{that}
\keyword{the}
\keyword{time}
\keyword{to}
\keyword{uses}
\keyword{will}
