% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{all_combinationsProcessing}
\alias{all_combinationsProcessing}
\title{Function to process all combinations
#' @keywords experimental}
\usage{
all_combinationsProcessing(
  edgeD,
  vertexD,
  no_samplesOut = 2,
  colorNodesTable,
  outputName = "CoNIOutput",
  outputDir = "./CoNIOutput"
)
}
\description{
Function to process all combinations
#' @keywords experimental
}
