% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{assign_colorsAnnotation}
\alias{assign_colorsAnnotation}
\title{AnnotationDf - Assing Colors Class}
\usage{
assign_colorsAnnotation(AnnotationDf, col = "Class")
}
\arguments{
\item{AnnotationDf}{Annotation data frame that contains a factor variable to use to assign colors}

\item{col}{Column with factor variable that will be used to assign colors}
}
\description{
This function assigns two color columns (color name and rgb) to an annotation data frame according to a column named 'Class' or 'class'
}
