% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{barplot_VertexsPerEdgeFeature}
\alias{barplot_VertexsPerEdgeFeature}
\title{Number Vertex features per class for every shared edge feature}
\usage{
barplot_VertexsPerEdgeFeature(
  SplitFile,
  title = "Vertex Features per class",
  AnnotationWithColors,
  ggrepelL = TRUE,
  xlb = "Gene",
  szggrepel = 2.5,
  szTitle = 12,
  szaxisTxt = 12,
  szaxisTitle = 12,
  szLegendText = 10,
  szLegendKey = 1
)
}
\value{
ggplot object for a barplot depicting the number of vertex features per class for every edge feature
}
\description{
Internal use. This function creates a barplot depicting the number of vertex features per class for every edge feature. To use this function one has to split first the file (or not if it is small) with the funciton chunk2
}
\keyword{internal}
