% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{Calculate_OverlapFrecuencyPerPair_allTreat}
\alias{Calculate_OverlapFrecuencyPerPair_allTreat}
\title{Overlap frequency}
\usage{
Calculate_OverlapFrecuencyPerPair_allTreat(Directory = "./")
}
\description{
Calculate the overlap frequency per triplet between all treatments
}
\keyword{experimental}
\keyword{internal}
