% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{check_outputDir}
\alias{check_outputDir}
\title{Output directory}
\usage{
check_outputDir(outputDir, verb)
}
\value{
No return value, called for side effects
}
\description{
Internal use. This function checks if the output directory exists and if it does not, it will create it
}
\keyword{internal}
