% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{check_previous}
\alias{check_previous}
\title{Check previous files}
\usage{
check_previous(del, iteration, outDir, verb)
}
\value{
No return value, called for side effects
}
\description{
Internal use. This function checks previous files and deletes according to the User input option
Requires some modifications...
}
\keyword{internal}
