% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{checkInputParameters}
\alias{checkInputParameters}
\title{Check input parameters}
\usage{
checkInputParameters(ParaList)
}
\description{
Internal use. Function  to check if input parameters are of the right class
}
\keyword{internal}
