% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{chunk2}
\alias{chunk2}
\title{Split function}
\usage{
chunk2(x, n)
}
\value{
A list of character vectors, each vector contains n edge features
}
\description{
Internal use. Function to split the EdgeFeatures in smaller groups
}
\keyword{internal}
