% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{compare_sampleNames}
\alias{compare_sampleNames}
\title{Compare sample names}
\usage{
compare_sampleNames(df1, df2)
}
\value{
No return value, called for side effects
}
\description{
Internal use. This function compares the sample names between the two datasets provided. If names do not match CoNI stops and outputs an error message.
}
\keyword{internal}
