% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{Compare_Triplets}
\alias{Compare_Triplets}
\title{Compare triplets}
\usage{
Compare_Triplets(
  Treat1_path,
  Treat2_path,
  OutputName = "Shared_Genes_and_Edges_Treat1vsTreat2.csv"
)
}
\arguments{
\item{Treat1_path}{TableForNetwork_file1 (file generated by CoNI) with path for Treatment 1}

\item{Treat2_path}{TableForNetwork_file2 (file generated by CoNI) with path for Treatment 2}

\item{OutputName}{Output file name with path}
}
\value{
A data.frame with the shared triplets (vertex1 vertex2 edge_feature) between two CoNI runs
}
\description{
Compare vertexFeature-vertexFeature-edgeFeature between two treatments, that is, find the shared triplets between two different CoNI runs.
}
\examples{
#For an example see the vignette
}
