% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{Compare_VertexClasses_sharedEdgeFeatures}
\alias{Compare_VertexClasses_sharedEdgeFeatures}
\title{Table VertexClass pairs of shared Edge Features}
\usage{
Compare_VertexClasses_sharedEdgeFeatures(
  Treat1_path,
  Treat2_path,
  OutputName = "Shared_Genes_and_Edges_Treat1vsTreat2.csv",
  Treat1Name = "Treat1",
  Treat2Name = "Treat2"
)
}
\arguments{
\item{Treat1_path}{TableForNetwork_file (file generated by CoNI) with path of Treatment 1}

\item{Treat2_path}{TableForNetwork_file (file generated by CoNI) with path of Treatment 2}

\item{OutputName}{Output file name with path}

\item{Treat1Name}{Name of treatment one, default Treat1}

\item{Treat2Name}{Name of treatment one, default Treat2}
}
\value{
A data.frame with all possible vertex-class pairs and their numbers per edge-feature and treatment.
}
\description{
Compare VertexClass pairs of the shared Edge Features of two treatments (e.g., lipid-class-pairs per shared gene)
}
\examples{
#For an example see the vignette
}
