% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{countClassPerEdgeFeature}
\alias{countClassPerEdgeFeature}
\title{Number lipid features per class}
\usage{
countClassPerEdgeFeature(ResTable, treatment = "Chow")
}
\description{
This function counts for every edge feature the number of (lipids/)features per class. It disregards the number of (lipid) pairs...
}
\keyword{internal}
