% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{create_GlobalBarplot}
\alias{create_GlobalBarplot}
\title{Vertex-class pairs profile of shared features between treatments}
\usage{
create_GlobalBarplot(
  CompTreatTable,
  treat1 = "Treatment1",
  treat2 = "Treatment2",
  factorOrder = NULL,
  col1 = "red",
  col2 = "blue",
  maxpairs = 1,
  nudgey = 0.5,
  nudgex = 0.5,
  xlb = "Vertex-Class Pairs",
  ylb = "Number of pairs",
  szggrepel = 3.5,
  szaxisTxt = 12,
  szaxisTitle = 12
)
}
\arguments{
\item{CompTreatTable}{Output of Compare_VertexClasses_sharedEdgeFeatures}

\item{treat1}{Name of treatment one, default Treatment1. It should match the column names of the output of Compare_VertexClasses_sharedEdgeFeatures}

\item{treat2}{Name of treatment one, default Treatment2. It should match the column names of the output of Compare_VertexClasses_sharedEdgeFeatures}

\item{factorOrder}{A list specifying the order of the treatments.}

\item{col1}{Color for Treatment 1}

\item{col2}{Color for Treatment 2}

\item{maxpairs}{If number of class-vertex-pairs > maxpairs, display number pairs on top of bar}

\item{szggrepel}{Size ggrepel labels}

\item{szaxisTxt}{Size axis text}

\item{szaxisTitle}{Size axis title}

\item{xlab}{Name for x-axis}

\item{ylab}{Name for the y-axis}
}
\value{
A barplot showing the vertex-class pairs profile of all shared edge features between treatments
}
\description{
This function will create a barplot from the output of Compare_VertexClasses_sharedEdgeFeatures using all shared Edge Features (e.g., genes).
}
\examples{
create_GlobalBarplot(CompTreatTable = LCP_HFDvsChow,
                     treat1 = "HFD",
                     treat2 = "Chow",
                     factorOrder = c("HFD","Chow"),
                     maxpairs = 1,xlb = "Metabolite-class-pairs")
}
