% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{createBipartiteGraph}
\alias{createBipartiteGraph}
\title{Create Bipartite Network}
\usage{
createBipartiteGraph(TableNetworkPath, colorVertexTable, incidenceMatrix = F)
}
\arguments{
\item{colorVertexTable}{Table specifying the colors for the vertex features. The first column should contain the names matching the features of the vertex Data and another column should specify the colors (column name: Colors).}

\item{incidenceMatrix}{logical. If TRUE it returns a hypergraph incidence matrix instead of a bipartite graph}

\item{TableNetwork}{TableForNetwork_file (file generated by CoNI) with path}
}
\value{
This function returns a bipartite graph (igraph object). Basic network statistics are included in the network, but ignoring that is a bipartite graph
}
\description{
This function creates a simple bipartite graph, it shows the driver and linked features as nodes.
}
\examples{
BipartiteGraph_Chow<-createBipartiteGraph("./Chow_example/TableForNetwork_Chow.csv",
                                          MetColorTable)
plot(BipartiteGraph_Chow) #Plot to screen
write.graph(BipartiteTest,file="Chow_example/BipartiteNetwork.graphml",format="graphml") #Save as graphml
}
