% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{createBipartiteTable}
\alias{createBipartiteTable}
\title{Bipartite Table}
\usage{
createBipartiteTable(CoNINetworkTable)
}
\description{
This function creates a table that is used to create a simple bipartite network
}
\keyword{internal}
