% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{delIntFiles}
\alias{delIntFiles}
\title{Delete intermediary files}
\usage{
delIntFiles(del, outDir)
}
\value{
No return value, called for side effects
}
\description{
Internal use. This function deletes intermediary files.
}
\keyword{internal}
