% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{do_objectsExist}
\alias{do_objectsExist}
\title{Check if files exist}
\usage{
do_objectsExist(gene_exp, norm_metabo_dat, verb)
}
\value{
No return value, called for side effects
}
\description{
Internal use. This function tests if the input files exist, if they do not it will output an error and
end CoNI
}
\keyword{internal}
