% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{find_localConfoundingFeatures}
\alias{find_localConfoundingFeatures}
\title{Find local confounding features}
\usage{
find_localConfoundingFeatures(ResultsCoNI, network, padjust = TRUE)
}
\arguments{
\item{ResultsCoNI}{The output of CoNI (after p-adjustment)}

\item{network}{Network created with the function generate_network}

\item{padjust}{logical. Filter output based on adjusted p values}
}
\value{
The function returns a table with the results of the binomial tests. Significant results correspond to local Confounding features
}
\description{
This function applies for a selected subnetwork a binomial test using the frequency of appearance of an edge feature and the total number of edge features. The probability corresponds to 1/n_df, where n_df corresponds to the total number of edge features in the network.
The selected subnetwork corresponds to the second level neighborhood of a specific node. The test is applied to all possible second level neighborhoods in the network.
}
