% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{generate_network_2}
\alias{generate_network_2}
\title{Create network}
\usage{
generate_network_2(ResultsCoNI, colorNodesTable)
}
\arguments{
\item{ResultsCoNI}{Output of CoNI}
}
\description{
This function creates a network using as input the output of CoNI and a table specifying the colors for the nodes (first column are node names)
}
\keyword{internal}
\keyword{not}
\keyword{used}
