% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{get_lowvarFeatures}
\alias{get_lowvarFeatures}
\title{Low variance features}
\usage{
get_lowvarFeatures(df)
}
\value{
A data.frame where low variance features were removed
}
\description{
Internal use. Get low variance features.
}
\keyword{internal}
