% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{get_ridNA}
\alias{get_ridNA}
\title{Get rid NA}
\usage{
get_ridNA(df, type = "column")
}
\description{
Internal use. This function gets rid of columns and/or rows with at least one NA
}
\keyword{internal}
