% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{getcolor}
\alias{getcolor}
\title{Get class rgb color
This function gets the class rgb color of a specific (lipid) class}
\usage{
getcolor(ClassM, tableColor)
}
\description{
Get class rgb color
This function gets the class rgb color of a specific (lipid) class
}
\keyword{internal}
