% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{getstackedGlobalBarplot_and_Grid}
\alias{getstackedGlobalBarplot_and_Grid}
\title{Stacked Global Barplot Side-by-side (two treatments)}
\usage{
getstackedGlobalBarplot_and_Grid(
  CompTreatTable,
  Treat1,
  Treat2,
  Annotation,
  ggrep = T,
  max_pairsLegend = 1,
  force = 0.1,
  xlb = "Vertex-Class Pairs",
  mx.overlaps = Inf,
  szggrepel = 6,
  ...
)
}
\arguments{
\item{CompTreatTable}{Output of Compare_VertexClasses_sharedEdgeFeatures}

\item{Treat1}{Name treatment 1 as in table CompTreatTable}

\item{Treat2}{Name treatment 2 as in table CompTreatTable}

\item{ggrep}{logical. If TRUE includes ggrepel labels for every bar}

\item{max_pairsLegend}{If number of Edge Features >= max_pairsLegend, display number of Edge Features}

\item{mx.overlaps}{Max number of overlaps ggrepel}

\item{szggrepel}{Size ggrepel labels}

\item{xlab}{Name for x-axis}
}
\value{
A stacked barplot showing the vertex-class pairs profile of all shared edge features between two treatments (one bar plot per treatment). Every bar consists of multiple genes that are depicted with different colors
}
\description{
This function will create a stacked barplot from the output of Compare_VertexClasses_sharedEdgeFeatures using all shared Edge Features (e.g., genes) between two treatments. Results of both Treatments are side by side for better comparison.
}
\examples{
HFD_vs_Chow_stackedBarplot<-getstackedGlobalBarplot_and_Grid(CompTreatTable = LCP_HFDvsChow,
                                                             Treat1 = "HFD",
                                                             Treat2 = "Chow",
                                                             Annotation = MetColorTable,
                                                             xlb = "Metabolite-class-pairs")
plot(HFD_vs_Chow_stackedBarplot) #Plot to screen
ggsave(HFD_vs_Chow_stackedBarplot,file="./HFD_vs_Chow_ClassProfile.svg") #or save it
}
