% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{getvertexes_edgeFeature}
\alias{getvertexes_edgeFeature}
\title{Get vertexes for edge feature}
\usage{
getvertexes_edgeFeature(edgeFeature, CoNIResults)
}
\value{
A character vector with the vertexes connected to a given edge feature
}
\description{
Get vertexes for edge feature
}
\keyword{internal}
