% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{HFD_GeneExpData}
\alias{HFD_GeneExpData}
\title{HFD gene expression data}
\description{
HFD gene expression data
}
\references{
\doi{10.1016/j.molmet.2021.101295}
}
\author{
Jose Manuel Monroy \email{nolozz@gmail.com}
}
\keyword{data}
