% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{labels2colors_2}
\alias{labels2colors_2}
\title{Labels to colors}
\usage{
labels2colors_2(
  labels,
  zeroIsGrey = TRUE,
  colorSeq = NULL,
  naColor = "grey",
  commonColorCode = TRUE
)
}
\value{
A character vector with colors
}
\description{
Internal use. This function is modified version of labels2colors of WGCNA and the internet
}
\keyword{internal}
