% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{MetaboliteAnnotation}
\alias{MetaboliteAnnotation}
\title{Metabolite Annotation}
\description{
Metabolite Annotation
}
\references{
\doi{10.1016/j.molmet.2021.101295}
}
\author{
Jose Manuel Monroy \email{nolozz@gmail.com}
}
\keyword{data}
