% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{obtain_groupcolors}
\alias{obtain_groupcolors}
\title{Get colors}
\usage{
obtain_groupcolors(Annotation)
}
\description{
This function gets the rgb colors of every (lipid) class and names them according to the class
}
\keyword{internal}
