% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{plotPcorvsCor}
\alias{plotPcorvsCor}
\title{Correlation vs Partial correlation}
\usage{
plotPcorvsCor(
  ResultsCoNI,
  edgeFeature,
  vertexD,
  edgeD,
  label_edgeFeature = "Gene",
  vertexFeatures = NULL,
  plot_to_screen = T,
  outputDir = "./",
  file_name = paste0(outputDir, edgeFeature, ".pdf"),
  height = 10,
  width = 8
)
}
\arguments{
\item{ResultsCoNI}{The significant results generated by CoNI}

\item{edgeFeature}{The edge feature to explore e.g. Fabp2 (for a gene)}

\item{vertexD}{Vertex data that was given as input to CoNI}

\item{edgeD}{Edge data that was given as input to CoNI}

\item{label_edgeFeature}{Name for plot title e.g. Gene or Protein}

\item{vertexFeatures}{The vertex features to include as a list. If not specified all metabolites available in combination with the edgeFeature will be used}

\item{plot_to_screen}{logical. If TRUE plots will be outputted to the plotting screen}

\item{file_name}{File name to save the plots}

\item{height}{height of the plotting area for the saved file}

\item{width}{width of the plotting are for the saved file}
}
\description{
This function generates a scatter plot with two regression lines. The slopes of the regression lines correspond to the correlation and partial correlation coefficients (blue for cor and red for pcor)
}
\examples{

#Plot gene of interest, included in the results of CoNI
plotPcorvsCor(ResultsCoNI = CoNIResults,edgeFeature = "Hpgd",
              vertexD = hfd_metabo,edgeD = hfd_gene,
              label_edgeFeature = "Gene",
              plot_to_screen = T,height = 10)

}
