% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{sig_correlation2Dfs}
\alias{sig_correlation2Dfs}
\title{Significant correlations 2 Df}
\usage{
sig_correlation2Dfs(metabolite_data, gene_expression)
}
\description{
Internal use. This function input are two data frames (e.g. metabolites and genes). It calculates the correlation matrix and creates a table with only significant pairs. No correction for multiple testing is done
}
\keyword{internal}
