% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{split_df}
\alias{split_df}
\title{Split dataset}
\usage{
split_df(AbundantDF, numberSplitDF_2 = 2)
}
\value{
A list of data.frames
}
\description{
Internal use. This function wills split the 'big' omics data into smaller data frames to improve computations, to avoid running out of memory.
}
\keyword{internal}
