% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{tableLCFs_VFs}
\alias{tableLCFs_VFs}
\title{Table local controlling edge features and vertex pairs}
\usage{
tableLCFs_VFs(CoNIResults, LCFs)
}
\arguments{
\item{CoNIResults}{The output of CoNI (after p-adjustment)}

\item{LCFs}{Local controlling edge features as a vector}
}
\value{
A data.frane of local controlling edge features and their respective vertex pairs, and unique vertexes.
}
\description{
This function creates a table of the local controlling edge features
}
\examples{
#Load CoNI results
data(CoNIResultsHFDToy)
#Note: arbitrary list of genes, not Local controlling features
tableLCFs_VFs(CoNIResultsHFDToy, c("Lilr4b","Rps12"))
}
