% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{VertexClassesSharedGenes_HFDvsChow}
\alias{VertexClassesSharedGenes_HFDvsChow}
\title{Toy data comparison treatments}
\description{
Toy data comparison Vertex classes shared edge featuresHFD vs Chow
}
\references{
\doi{10.1016/j.molmet.2021.101295}
}
\author{
Jose Manuel Monroy \email{nolozz@gmail.com}
}
\keyword{data}
\keyword{toydata}
