% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{countClassPerEdgeFeature}
\alias{countClassPerEdgeFeature}
\title{Number lipid features per class}
\usage{
countClassPerEdgeFeature(ResTable, treatment = "Chow")
}
\value{
A data.frame with the number of vertex features per class and edge feature
}
\description{
Internal use. This function counts for every edge feature the number of vertex features (e.g. lipids) per class. It disregards the number of vertex pairs...
}
\keyword{internal}
