% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{createBipartiteGraph}
\alias{createBipartiteGraph}
\title{Bipartite Network}
\usage{
createBipartiteGraph(TableNetwork, colorVertexTable, incidenceMatrix = FALSE)
}
\arguments{
\item{TableNetwork}{TableForNetwork_file (file generated by CoNI) with path}

\item{colorVertexTable}{Table specifying the colors for the vertex features. The first column should contain the names matching the features of the vertex Data and another column should specify the colors (column name: Colors).}

\item{incidenceMatrix}{logical. If TRUE it returns a hypergraph incidence matrix instead of a bipartite graph}
}
\value{
An igraph object for a bipartite graph or a hypergraph incidence matrix to represent ResultsCoNI. Basic network statistics are included in the bipartite graph. See generate_network function for details or consult the igraph package
}
\description{
This function creates a simple bipartite graph, it shows the driver and linker features as nodes.
}
\examples{
#See vignette for an example
}
