% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{createBipartiteTable}
\alias{createBipartiteTable}
\title{Bipartite Table}
\usage{
createBipartiteTable(CoNINetworkTable)
}
\value{
A matrix containing two columns, one for vertexes and one for edge features. The matching happens if they are adjacent to one another
}
\description{
Internal use. This function creates a table that is used to create a simple bipartite network
}
\keyword{internal}
