% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{getcolor}
\alias{getcolor}
\title{Get class rgb color}
\usage{
getcolor(ClassM, tableColor)
}
\value{
A character object, that corresponds to a color in hexadecimal format
}
\description{
Internal use. This function gets the class rgb color of a specific (lipid) class
}
\keyword{internal}
