% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{NetStats}
\alias{NetStats}
\title{Network Statistics}
\usage{
NetStats(Network)
}
\arguments{
\item{Network}{An Igraph network}
}
\value{
Returns network statistics. For more information on the statistics consult the igraph package.
}
\description{
This function calculates simple network statistics and returns them as a dataframe
}
\examples{
NetStats(HFDNetwork)
}
