% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{top_n_LF_byMagnitude}
\alias{top_n_LF_byMagnitude}
\title{Linker Features by magnitude of effect}
\usage{
top_n_LF_byMagnitude(ResultsCoNI, topn = 10)
}
\arguments{
\item{ResultsCoNI}{The output of CoNI}

\item{topn}{Top n number of features to output}
}
\value{
Returns a data.frame, a filtered version of ResultsCoNI, showing the top n features
with the strongest effect, that is, the highest difference between the partial correlation and correlation coefficient.
}
\description{
This function outputs the linker features with the strongest effect on the correlation of the vertex features
}
\examples{
data(CoNIResultsHFDToy)
Top10HFD<-top_n_LF_byMagnitude(CoNIResults_HFD,topn = 10)
}
