% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNI_functions.R
\name{writeTable}
\alias{writeTable}
\title{Write table}
\usage{
writeTable(results_write, num_cores, outputDir, iteration)
}
\value{
Returns NA if it fails to write file to memory, otherwise no return value
}
\description{
Internal use. This function tries to write a table with fread, if it fails it returns "NA"
}
\keyword{internal}
